ProjectDatabase
It performs all CRUD, but when you try to delete something, it does not reflect immediately on the layout, app restart is needed to make it reflect on the UI/UX. Still trying to fix it. Made a refresh feature to try to solve it but it's still not working properly.